+++ esp-hal-common/src/spi/slave.rs
@@ -1,11 +1,13 @@
-//! # Serial Peripheral Interface, slave mode
+//! # Serial Peripheral Interface - Slave Mode
 //!
 //! ## Overview
+//!
 //! There are multiple ways to use SPI, depending on your needs. Regardless of
 //! which way you choose, you must first create an SPI instance with
 //! [`Spi::new`].
 //!
 //! ## Example
+//!
 //! ```rust
 //! let io = IO::new(peripherals.GPIO, peripherals.IO_MUX);
 //! let sclk = io.pins.gpio12;
@@ -13,7 +15,7 @@
 //! let mosi = io.pins.gpio13;
 //! let cs = io.pins.gpio10;
 //!
-//! let mut spi = hal::spi_slave::Spi::new(peripherals.SPI2, sclk, mosi, miso, cs, SpiMode::Mode0);
+//! let mut spi = hal::spi::slave::Spi::new(peripherals.SPI2, sclk, mosi, miso, cs, SpiMode::Mode0);
 //! ```
 //!
 //! There are several options for working with the SPI peripheral in slave mode,
@@ -57,6 +59,7 @@
 
 use core::marker::PhantomData;
 
+use super::{Error, FullDuplexMode, SpiMode};
 use crate::{
     dma::{DmaPeripheral, Rx, Tx},
     gpio::{InputPin, InputSignal, OutputPin, OutputSignal},
@@ -66,7 +69,6 @@ use crate::{
 };
 
 const MAX_DMA_SIZE: usize = 32768 - 32;
-pub use crate::spi::{Error, FullDuplexMode, SpiMode};
 
 /// SPI peripheral driver
 pub struct Spi<'d, T, M> {
@@ -128,12 +130,8 @@ pub mod dma {
 
     use embedded_dma::{ReadBuffer, WriteBuffer};
 
-    #[cfg(any(esp32, esp32s2, esp32s3))]
-    use super::Spi3Instance;
-    #[allow(unused_imports)]
-    use super::SpiMode;
-    use super::{FullDuplexMode, Instance, InstanceDma, Spi, Spi2Instance, MAX_DMA_SIZE};
-    #[cfg(any(esp32, esp32s2, esp32s3))]
+    use super::*;
+    #[cfg(spi3)]
     use crate::dma::Spi3Peripheral;
     use crate::{
         dma::{
@@ -159,7 +157,7 @@ pub mod dma {
         fn with_dma(self, channel: Channel<'d, C>) -> SpiDma<'d, T, C>;
     }
 
-    #[cfg(any(esp32, esp32s2, esp32s3))]
+    #[cfg(spi3)]
     pub trait WithDmaSpi3<'d, T, C>
     where
         T: Instance + Spi3Instance,
@@ -193,7 +191,7 @@ pub mod dma {
         }
     }
 
-    #[cfg(any(esp32, esp32s2, esp32s3))]
+    #[cfg(spi3)]
     impl<'d, T, C> WithDmaSpi3<'d, T, C> for Spi<'d, T, FullDuplexMode>
     where
         T: Instance + Spi3Instance,
@@ -388,14 +386,14 @@ pub mod dma {
         pub fn dma_write<TXBUF>(
             mut self,
             words: TXBUF,
-        ) -> Result<SpiDmaTransfer<'d, T, C, TXBUF>, super::Error>
+        ) -> Result<SpiDmaTransfer<'d, T, C, TXBUF>, Error>
         where
             TXBUF: ReadBuffer<Word = u8>,
         {
             let (ptr, len) = unsafe { words.read_buffer() };
 
             if len > MAX_DMA_SIZE {
-                return Err(super::Error::MaxDmaTransferSizeExceeded);
+                return Err(Error::MaxDmaTransferSizeExceeded);
             }
 
             self.spi
@@ -416,14 +414,14 @@ pub mod dma {
         pub fn dma_read<RXBUF>(
             mut self,
             mut words: RXBUF,
-        ) -> Result<SpiDmaTransfer<'d, T, C, RXBUF>, super::Error>
+        ) -> Result<SpiDmaTransfer<'d, T, C, RXBUF>, Error>
         where
             RXBUF: WriteBuffer<Word = u8>,
         {
             let (ptr, len) = unsafe { words.write_buffer() };
 
             if len > MAX_DMA_SIZE {
-                return Err(super::Error::MaxDmaTransferSizeExceeded);
+                return Err(Error::MaxDmaTransferSizeExceeded);
             }
 
             self.spi
@@ -446,7 +444,7 @@ pub mod dma {
             mut self,
             words: TXBUF,
             mut read_buffer: RXBUF,
-        ) -> Result<SpiDmaTransferRxTx<'d, T, C, RXBUF, TXBUF>, super::Error>
+        ) -> Result<SpiDmaTransferRxTx<'d, T, C, RXBUF, TXBUF>, Error>
         where
             TXBUF: ReadBuffer<Word = u8>,
             RXBUF: WriteBuffer<Word = u8>,
@@ -455,7 +453,7 @@ pub mod dma {
             let (read_ptr, read_len) = unsafe { read_buffer.write_buffer() };
 
             if write_len > MAX_DMA_SIZE || read_len > MAX_DMA_SIZE {
-                return Err(super::Error::MaxDmaTransferSizeExceeded);
+                return Err(Error::MaxDmaTransferSizeExceeded);
             }
 
             self.spi
@@ -582,7 +580,7 @@ where
     fn dma_peripheral(&self) -> DmaPeripheral {
         match self.spi_num() {
             2 => DmaPeripheral::Spi2,
-            #[cfg(any(esp32, esp32s2, esp32s3))]
+            #[cfg(spi3)]
             3 => DmaPeripheral::Spi3,
             _ => panic!("Illegal SPI instance"),
         }
@@ -688,7 +686,7 @@ where
 {
 }
 
-#[cfg(any(esp32, esp32s2, esp32s3))]
+#[cfg(spi3)]
 impl<TX, RX> InstanceDma<TX, RX> for crate::peripherals::SPI3
 where
     TX: Tx,
@@ -1080,10 +1078,10 @@ impl Instance for crate::peripherals::SPI3 {
 
 pub trait Spi2Instance {}
 
-#[cfg(any(esp32, esp32s2, esp32s3))]
+#[cfg(spi3)]
 pub trait Spi3Instance {}
 
 impl Spi2Instance for crate::peripherals::SPI2 {}
 
-#[cfg(any(esp32, esp32s2, esp32s3))]
+#[cfg(spi3)]
 impl Spi3Instance for crate::peripherals::SPI3 {}