+++ esp32h2-hal/examples/spi_halfduplex_read_manufacturer_id.rs
@@ -22,7 +22,11 @@ use esp32h2_hal::{
     gpio::IO,
     peripherals::Peripherals,
     prelude::*,
-    spi::{Address, Command, HalfDuplexReadWrite, Spi, SpiDataMode, SpiMode},
+    spi::{
+        master::{Address, Command, HalfDuplexReadWrite, Spi},
+        SpiDataMode,
+        SpiMode,
+    },
     Delay,
 };
 use esp_backtrace as _;