+++ esp32s2-hal/examples/spi_eh1_device_loopback.rs
@@ -24,7 +24,10 @@ use esp32s2_hal::{
     gpio::IO,
     peripherals::Peripherals,
     prelude::*,
-    spi::{Spi, SpiBusController, SpiMode},
+    spi::{
+        master::{Spi, SpiBusController},
+        SpiMode,
+    },
     Delay,
 };
 use esp_backtrace as _;